requirements.txt
adafruit-ampy==1.0.7
Click==7.0
docopt==0.6.2
esp32-machine-emulator==1.1.3
pyserial==3.4
python-dotenv==0.10.3
